LPS-199470 Inline SiteMembershipPolicyFactoryUtil logic into SiteMembershipPolicyUtil
# breaking_change_report

## What portal-impl/src/com/liferay/portal/security/membershippolicy/SiteMembershipPolicyFactoryUtil.java

SiteMembershipPolicyFactoryUtil has been removed

## Why
It is for task 'LPS-190426', we would like to merge SiteMembershipPolicyFactoryUtil logic into SiteMembershipPolicyUtil

## Alternatives
We can use the static method in SiteMembershipPolicyUtil instead

----
